Add files via upload
passed all programflow tests (2 tests in the folder) , simpleFunction.vm test
IMPORTANT - when I tested I used the CPUemulator. in the program flow tests for example, I tested our ASM output vs the BasicLoop.tst file. Notice there is also a BasicLoopVME.tst file, which I did not understand what it means, it is some kind of wrapper for the whole process or something - SHIRI its your job to make sure we pass them also!
Need to test the rest of the tests in function call folder.
TODO: Check Sys.init function call, I supplied the logic but did not understand how to test - For you to find out
Fix problems from project 7 (as we got feedback in Moodle)
Make sure we did not miss anything else that is needed for this project..